python_cs/test/pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "^4.1.5"
pandas = "^2.2.1"
scipy = "^1.12.0"
openai = "^1.14.3"
langchain-community = "^0.0.29"
tiktoken = "^0.6.0"
langchain-openai = "^0.1.1"
langchainhub = "^0.1.15"
chromadb = "^0.4.24"
langchain = "^0.1.13"
pillow = "^10.3.0"
cython = "^3.0.10"
pypng = "^0.20220715.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
